Subject: [PATCH] irqchip: or1k-pic: Fix interrupt system.
Usually, hardware implicitly acknowledges interruts when
reading them. However, if this is not the case, the IRQ
gets fired over and over again in the current implementation.

This patch uses the right mask acknowledge function to handle the
aforementioned situation on or1k processors that interact with
such kind of hardware.

Signed-off-by: Pedro H. Penna <pedrohenriquepenna@gmail.com>